api/project_cache/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
cryptography==45.0.7
distro==1.9.0
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.111.0
fastapi-cli==0.0.11
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
lxml==6.0.1
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
msal==1.33.0
openai==1.107.0
orjson==3.11.3
pycparser==2.23
pydantic==2.8.2
pydantic-settings==2.3.4
pydantic_core==2.20.1
Pygments==2.19.2
PyJWT==2.10.1
PyPDF2==3.0.1
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.2
requests==2.32.5
rich==14.1.0
rich-toolkit==0.15.1
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.30
starlette==0.37.2
tqdm==4.67.1
typer==0.17.4
typing-inspection==0.4.1
typing_extensions==4.15.0
ujson==5.11.0
urllib3==2.5.0
uvicorn==0.30.1
watchfiles==1.1.0
websockets==15.0.1